Refactor GcloudHandler into modular services
Decomposed the monolithic GcloudHandler class into domain-specific services:
- GcloudExecutor: Handles command execution and environment context.
- GcloudInstallService: Manages gcloud installation and version checks.
- GcloudAuthService: Handles authentication flows (User/ADC).
- GcloudProjectService: Manages project listing and setting.

GcloudHandler now acts as a facade, ensuring backward compatibility.
Updated tests to reflect the new internal structure.
